+++ modules/config.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2019 João Pedro Martins Neves - All Rights Reserved.
+ * Copyright (c) 2018-2020 João Pedro Martins Neves - All Rights Reserved.
  *
  * MergerJS (merger-js) is licensed under the MIT license, located in
  * the root of this project, under the name "LICENSE.md".