+++ tests/data/headerFiles/relativeDirPaths.header.js
@@ -1,2 +1,10 @@
-﻿// @import<<DIR '../node_modules'
+﻿/*
+ * Copyright (c) 2018-2020 João Pedro Martins Neves - All Rights Reserved.
+ *
+ * MergerJS (merger-js) is licensed under the MIT license, located in
+ * the root of this project, under the name "LICENSE.md".
+ *
+ */
+
+// @import<<DIR '../node_modules'
 // @import<<directory '../node_modules/randomDir1'